Logs file problems fixed
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@293 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/cgi.cpp
@@ -451,8 +451,8 @@ void buildCGIEnvironmentString(httpThreadContext* td,char *cgiEnvString,int proc
 #endif
 
 	strcat(cgiEnvString,"\r\0\0\0\0\0");
-	int max=strlen(cgiEnvString);
-	for(int i=0;i<max;i++)
+	size_t max=strlen(cgiEnvString);
+	for(size_t i=0;i<max;i++)
 		if(cgiEnvString[i]=='\r')
 			cgiEnvString[i]='\0';
 }
